[safe-vibe-coding] Notifications of sandbox denials (#13)
We want to be able to be notified when the AppArmor sandbox produces a
denial.

This change now allows us to get notified of what path produced the
denial. It does so by inspecting the `common_lsm_audit` kernel function
and checking if the cgroup of the program that caused the crash matches
the cgroup that was requested.